requirements.txt
flask==2.0.1
gunicorn==20.1.0
requests==2.26.0
python-dotenv==0.19.1
google-api-python-client==2.35.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-generativeai==0.3.0
openai==1.0.0
python-telegram-bot==13.7
